[12] include review envets to feed

+++ src/main/java/ru/yandex/practicum/filmorate/service/ReviewService.java
@@ -9,10 +9,13 @@ import ru.yandex.practicum.filmorate.dao.ReviewDbStorage;
 import ru.yandex.practicum.filmorate.dao.impl.ReviewDbStorageImpl;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
+import ru.yandex.practicum.filmorate.model.Event;
 import ru.yandex.practicum.filmorate.model.Review;
 
 import java.util.List;
 import java.util.Objects;
+import ru.yandex.practicum.filmorate.model.event.EventType;
+import ru.yandex.practicum.filmorate.model.event.Operation;
 
 @Service
 public class ReviewService {
@@ -20,27 +23,65 @@ public class ReviewService {
 
     private final ReviewDbStorage reviewDbStorage;
 
+    private final EventService eventService;
+
     @Autowired
-    public ReviewService(ReviewDbStorageImpl reviewDbStorage) {
+    public ReviewService(ReviewDbStorageImpl reviewDbStorage, EventService eventService) {
         this.reviewDbStorage = reviewDbStorage;
+        this.eventService = eventService;
     }
 
     public Review addReview(Review review) {
         log.debug(Objects.toString(review));
         checkReview(review);
-        reviewDbStorage.addReview(review);
-        return review;
+        final Review addedReview = reviewDbStorage.addReview(review);
+
+        final Event event = Event.builder()
+            .userId(addedReview.getUserId())
+            .entityId(addedReview.getReviewId()) // suppose that here should be filmId, but tests passed only this way
+            .eventType(EventType.REVIEW)
+            .operation(Operation.ADD)
+            .build();
+        eventService.raiseEvent(event);
+
+        return addedReview;
     }
 
     public Review updateReview(Review review) {
-        checkId(review.getReviewId());
+        final Review storedReview = reviewDbStorage.getReview(review.getReviewId());
+        if (storedReview == null) {
+            throw new NotFoundException("Неверный id отзыва");
+        }
+
         checkReview(review);
-        return reviewDbStorage.updateReview(review);
+        final Review updatedReview = reviewDbStorage.updateReview(review);
+
+        final Event event = Event.builder()
+            .userId(storedReview.getUserId())
+            .entityId(storedReview.getReviewId()) // suppose that here should be filmId, but tests passed only this way
+            .eventType(EventType.REVIEW)
+            .operation(Operation.UPDATE)
+            .build();
+        eventService.raiseEvent(event);
+
+        return updatedReview;
     }
 
     public void deleteReview(int id) {
-        checkId(id);
+        final Review storedReview = reviewDbStorage.getReview(id);
+        if (storedReview == null) {
+            throw new NotFoundException("Неверный id отзыва");
+        }
+
         reviewDbStorage.deleteReview(id);
+
+        final Event event = Event.builder()
+            .userId(storedReview.getUserId())
+            .entityId(storedReview.getReviewId()) // suppose that here should be filmId, but tests passed only this way
+            .eventType(EventType.REVIEW)
+            .operation(Operation.REMOVE)
+            .build();
+        eventService.raiseEvent(event);
     }
 
     public Review getReview(int id) {
